01_etl/requirements.txt
flake8==4.0
python-dotenv==0.20.0
psycopg2-binary==2.9.3
pytest==7.1.1
elasticsearch==8.1.3
backoff==2.0.1
redis=4.2.2